darlingCms_0.0_dev|core|classes: The \DarlingCms\classes\crud\json() class now keeps a registry of all the data it stores. Defined new method updateRegistry() which is responsible for updating the registry whenever data is saved or deleted. Defined new method reloadRegistry() which is responsible for re-loading the registry into the $registry property, reloadRegistry() is called whenever updateRegistry() is called so the $registry property stays in sync with the stored registry. Defined new method getRegistry() which returns the registry as an associative array where keys are storage ids and values are safe ids. Defined new method getStoragePath() which returns the full storage path to where the \DarlingCms\classes\crud\json() class stores the data it manages. The \DarlingCms\classes\startup\darlingCmsStartup()'s __construct() method now expects all parameters be instances of the \DarlingCms\interfaces\startup\Istartup interface, the first and second parameters must still instances of the \DarlingCms\classes\startup\appStartup and \DarlingCms\classes\startup\themeStartup implementations of the \DarlingCms\interfaces\startup\Istartup interface, respectively. Corrected doc comments in ./core/classes/startup/darlingCmsStartup.php and ./core/classes/startup/themeStartup.php.

+++ core/classes/crud/json.php
@@ -8,7 +8,6 @@
 
 namespace DarlingCms\classes\crud;
 
-
 /**
  * Class json. Implementation of the \DarlingCms\abstractions\crud\Acrud() abstract
  * class that uses JSON files as it's storage mechanism.
@@ -17,13 +16,50 @@ namespace DarlingCms\classes\crud;
 class json extends \DarlingCms\abstractions\crud\Acrud
 {
     private $storagePath;
+    private $registry;
 
     /**
-     * json constructor. Initializes the $storagePath.
+     * json constructor. Initializes the $storagePath, and the $registry.
      */
     public function __construct()
     {
+        /* Initialize storage path. */
         $this->storagePath = trim(str_replace('core/classes/crud', '', __DIR__) . '.dcms/');
+        /* Make sure the stored registry exists. */
+        if ($this->read('registry') === false) {
+            /* If the stored registry does not exist, create it. */
+            $this->create('registry', array('registry' => $this->safeId('registry')));
+        }
+        /* Initialize $registry using stored registry. */
+        $this->registry = $this->read('registry');
+    }
+
+    /**
+     * Generates an id that is safe to use for storage from the $storageId.
+     * @param string $storageId The storage id to generate a safe id for.
+     * @return string An id generated from the $storageId that is safe to use for storage.
+     */
+    private function safeId(string $storageId)
+    {
+        return hash_hmac('sha256', $storageId, 'ads9fuj4ih98fyhudihf908ydfgv2goe7r87gwe');
+    }
+
+    /**
+     * Returns the registry as an associative array where keys are storage ids and values are safe ids.
+     * @return mixed The registry as an associative array.
+     */
+    public function getRegistry()
+    {
+        return $this->read('registry');
+    }
+
+    /**
+     * Returns the full storage path to where data is stored.
+     * @return string The storage path.
+     */
+    public function getStoragePath()
+    {
+        return $this->storagePath;
     }
 
     /**
@@ -42,6 +78,8 @@ class json extends \DarlingCms\abstractions\crud\Acrud
         $pathToJsonFile = $this->storagePath . $this->safeId($storageId) . '.json';
         switch ($mode) {
             case 'save':
+                $this->registry[$storageId] = $this->safeId($storageId);
+                $this->updateRegistry($storageId);
                 return (file_put_contents($pathToJsonFile, $data, LOCK_EX) > 0);
             case 'load':
                 if (file_exists($pathToJsonFile) === true) {
@@ -49,7 +87,12 @@ class json extends \DarlingCms\abstractions\crud\Acrud
                 }
                 return false;
             case 'delete':
-                return unlink($pathToJsonFile);
+                unset($this->registry[$storageId]);
+                $this->updateRegistry($storageId);
+                if (file_exists($pathToJsonFile) === true) {
+                    return unlink($pathToJsonFile);
+                }
+                return false;
             default:
                 error_log('Attempt to use invalid query mode in ' . __FILE__);
                 return false;
@@ -57,13 +100,31 @@ class json extends \DarlingCms\abstractions\crud\Acrud
     }
 
     /**
-     * Generates an id that is safe to use for storage from the $storageId.
-     * @param string $storageId The storage id to generate a safe id for.
-     * @return string An id generated from the $storageId that is safe to use for storage.
+     * Update the registry.
+     * @param string $storageId The $storageId of the data currently being processed.
+     * @return bool True if registry was updated, false otherwise.
      */
-    private function safeId(string $storageId)
+    private function updateRegistry(string $storageId)
     {
-        return hash_hmac('sha256', $storageId, 'ads9fuj4ih98fyhudihf908ydfgv2goe7r87gwe');
+        $status = false;
+        if ($storageId !== 'registry') {
+            $status = $this->update('registry', $this->registry);
+        }
+        $this->reloadRegistry();
+        return $status;
+    }
+
+    /**
+     * Re-load the registry into the $registry property. This method is called whenever
+     * the registry is updated, i.e., whenever updateRegistry() is called.
+     * @return bool True if registry was re-loaded, false otherwise.
+     */
+    private function reloadRegistry()
+    {
+        $origRegistry = $this->registry;
+        unset($this->registry);
+        $this->registry = $this->read('registry');
+        return ($this->registry !== $origRegistry);
     }
 
     /**
@@ -73,4 +134,5 @@ class json extends \DarlingCms\abstractions\crud\Acrud
     {
         return json_encode(base64_encode(serialize($data)));
     }
+
 }
